recipe_chatbot/recipe_rag_project/requirements.txt
# Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0

# LangChain
langchain==0.3.9
langchain-openai==0.2.9
langchain-chroma==0.1.4
langchain-core==0.3.21

# Vector DB
chromadb==0.5.23

# OpenAI
openai==1.55.3

# Optional: Development
pytest==8.3.4
black==24.10.0
ruff==0.8.2
